A lot of name refactoring and fixed expiration time
